location-module/requirements.txt
absl-py==1.3.0
arrow==1.2.3
astunparse==1.6.3
blis==0.7.9
cachetools==5.2.0
catalogue==1.0.2
certifi==2021.10.8
chardet==5.0.0
charset-normalizer==2.0.10
click==8.0.3
cymem==2.0.7
Deprecated==1.2.13
editdistance==0.6.1
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
en-core-web-lg==2.3.1
flatbuffers==22.10.26
gast==0.4.0
geographiclib==1.52
geopy==2.2.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.3
importlib-metadata==5.0.0
joblib==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
langdetect==1.0.9
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mediacloud-cliff==2.6.1
mordecai==2.1.0
murmurhash==1.0.9
nltk==3.6.7
numpy==1.23.4
oauthlib==3.2.2
opensextant==1.4.7
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
plac==1.1.3
preshed==3.0.8
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld3==0.22
pycountry==22.1.10
PyGeodesy==22.10.10
pymongo==4.0.1
pyparsing==3.0.6
pysolr==3.9.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.1.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
spacy==2.3.8
srsly==1.0.6
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
thinc==7.4.6
tqdm==4.62.3
typing-extensions==4.4.0
Unidecode==1.3.2
urllib3==1.26.7
wasabi==0.10.1
Werkzeug==2.2.2
wrapt==1.13.3
zipp==3.10.0
